backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.48.0
jinja2==3.1.4
pillow==10.4.0
aiosqlite==0.20.0
pydantic==2.9.2
python-multipart==0.0.12
yt-dlp>=2024.1.0
easyocr>=1.7.0
deep-translator>=1.11.0
langdetect>=1.0.9
